+++ frontend/src/services/config_service.rs
@@ -77,6 +77,11 @@ impl ConfigService {
                 for source in app_config.sources.sources.iter_mut() {
                     for target in source.targets.iter_mut() {
                         target.t_filter = get_filter(target.filter.as_str(), templates.as_ref()).ok();
+                        if let Some(sort) = target.sort.as_mut() {
+                            for channel in sort.rules.iter_mut() {
+                                channel.t_filter = get_filter(&channel.filter, templates.as_ref()).map_err(|e| error!("Failed to parse groups sort filter: {}", e)).ok();
+                            }
+                        }
                         for output in target.output.iter_mut() {
                             match output {
                                 TargetOutputDto::Xtream(o) =>