+++ shared/src/model/config/sort.rs
@@ -1,9 +1,11 @@
-use std::fmt::Display;
 use crate::error::{TuliproxError, TuliproxErrorKind};
-use crate::foundation::filter::{apply_templates_to_pattern, apply_templates_to_pattern_single};
+use crate::foundation::filter::{apply_templates_to_pattern, get_filter, Filter};
 use crate::model::{ItemField, PatternTemplate, TemplateValue};
-use crate::{info_err, handle_tuliprox_error_result_list};
+use crate::{handle_tuliprox_error_result_list, info_err, info_err_res};
 use regex::Regex;
+use serde::{Deserialize, Deserializer, Serialize};
+use std::fmt::{Display, Formatter};
+use std::str::FromStr;
 
 fn compile_regex_vec(patterns: Option<&Vec<String>>) -> Result<Option<Vec<Regex>>, TuliproxError> {
     patterns.as_ref()
@@ -38,76 +40,98 @@ impl Display for SortOrder {
     }
 }
 
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct ConfigSortGroupDto {
-    pub order: SortOrder,
-    #[serde(default)]
-    pub sequence: Option<Vec<String>>,
-    #[serde(skip)]
-    pub t_sequence: Option<Vec<Regex>>,
+
+#[derive(Serialize, Debug, Copy, Clone, Eq, PartialEq)]
+pub enum SortTarget {
+    #[serde(rename = "group")]
+    Group,
+    #[serde(rename = "channel")]
+    Channel,
 }
 
-impl PartialEq for ConfigSortGroupDto {
-    fn eq(&self, other: &Self) -> bool {
-        self.order == other.order
-            && self.sequence == other.sequence
+impl SortTarget {
+    const GROUP: &'static str = "group";
+    const CHANNEL: &'static str = "channel";
+
+    pub fn as_str(&self) -> &'static str {
+        match *self {
+            Self::Group => Self::GROUP,
+            Self::Channel => Self::CHANNEL,
+        }
     }
 }
 
-impl ConfigSortGroupDto {
-    pub fn prepare(&mut self, templates: Option<&Vec<PatternTemplate>>) -> Result<(), TuliproxError> {
-        let processed_sequence = match (&self.sequence, templates) {
-            (Some(seqs), Some(_templs)) => {
-                let mut result = Vec::new();
-                for s in seqs {
-                    match apply_templates_to_pattern(s, templates, true)? {
-                        TemplateValue::Single(val) => result.push(val),
-                        TemplateValue::Multi(vals) => result.extend(vals),
-                    }
-                }
-                Some(result)
-            }
-            (Some(seqs), None) => Some(seqs.clone()),
-            (None, _) => None,
-        };
+impl FromStr for SortTarget {
+    type Err = TuliproxError;
 
-        self.t_sequence = compile_regex_vec(processed_sequence.as_ref())?;
-        Ok(())
+    fn from_str(s: &str) -> Result<Self, TuliproxError> {
+        if s.eq_ignore_ascii_case(Self::GROUP) {
+            Ok(Self::Group)
+        } else if s.eq_ignore_ascii_case(Self::CHANNEL) {
+            Ok(Self::Channel)
+        } else {
+            info_err_res!("Unknown SortTarget: {}", s)
+        }
+    }
+}
+
+impl<'de> Deserialize<'de> for SortTarget {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s = String::deserialize(deserializer)?;
+        SortTarget::from_str(&s).map_err(serde::de::Error::custom)
+    }
+}
+
+impl Display for SortTarget {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}",
+               match self {
+                   Self::Group => Self::GROUP,
+                   Self::Channel => Self::CHANNEL,
+               }
+        )
     }
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 #[serde(deny_unknown_fields)]
-pub struct ConfigSortChannelDto {
-    // channel field
+pub struct ConfigSortRuleDto {
+    pub target: SortTarget,
+    // channel/Group field
     pub field: ItemField,
-    // match against group title
-    pub group_pattern: String,
     pub order: SortOrder,
     #[serde(default)]
     pub sequence: Option<Vec<String>>,
+    pub filter: String,
     #[serde(skip)]
     pub t_sequence: Option<Vec<Regex>>,
+    #[serde(skip)]
+    pub t_filter: Option<Filter>,
 }
 
-impl PartialEq for ConfigSortChannelDto {
+impl PartialEq for ConfigSortRuleDto {
     fn eq(&self, other: &Self) -> bool {
         self.field == other.field
-            && self.group_pattern == other.group_pattern
             && self.order == other.order
             && self.sequence == other.sequence
+            && self.filter == other.filter
     }
 }
 
 
-impl ConfigSortChannelDto {
+impl ConfigSortRuleDto {
     pub fn prepare(&mut self, templates: Option<&Vec<PatternTemplate>>) -> Result<(), TuliproxError> {
-        self.group_pattern = apply_templates_to_pattern_single(&self.group_pattern, templates)?;
-        // Compile group_pattern
-        Regex::new(&self.group_pattern).map_err(|err| {
-            info_err!("can't parse regex: {} {err}", &self.group_pattern)
-        })?;
+        if self.target == SortTarget::Group  {
+            if !matches!(self.field, ItemField::Group | ItemField::Title | ItemField::Name | ItemField::Caption) {
+                return info_err_res!("Group sorting can only be done on the Group field");
+            }
+             self.field = ItemField::Group; // hard coded because we only can't match a group until we can use PlaylistGroup with filter
+        }
+
+        self.t_filter = Some(get_filter(&self.filter, templates)?);
 
         // Transform sequence with templates if provided, otherwise use raw sequence
         let processed_sequence = match (&self.sequence, templates) {
@@ -137,20 +161,12 @@ pub struct ConfigSortDto {
     #[serde(default)]
     pub match_as_ascii: bool,
     #[serde(default)]
-    pub groups: Option<ConfigSortGroupDto>,
-    #[serde(default)]
-    pub channels: Option<Vec<ConfigSortChannelDto>>,
+    pub rules: Vec<ConfigSortRuleDto>,
 }
-
 
 impl ConfigSortDto {
     pub fn prepare(&mut self, templates: Option<&Vec<PatternTemplate>>) -> Result<(), TuliproxError> {
-        if let Some(group) = self.groups.as_mut() {
-            group.prepare(templates)?;
-        }
-        if let Some(channels) = self.channels.as_mut() {
-            handle_tuliprox_error_result_list!(TuliproxErrorKind::Info, channels.iter_mut().map(|csc| csc.prepare(templates)));
-        }
+        handle_tuliprox_error_result_list!(TuliproxErrorKind::Info, self.rules.iter_mut().map(|rule| rule.prepare(templates)));
         Ok(())
     }
 }